+++ src/assets/RightArrow.tsx
@@ -1,0 +1,25 @@
+/* eslint-disable react/no-unknown-property */
+import React from "react"
+function RightArrow() {
+  return (
+    <button className="svg-arrow svg-arrow-right">
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="25px"
+        height="50px"
+        viewBox="0 0 50 80"
+        xmlSpace="preserve"
+      >
+        <polyline
+          fill="none"
+          stroke="#FFFFFF"
+          strokeWidth="5"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          points="0.375,0.375 45.63,38.087 0.375,75.8"
+        />
+      </svg>
+    </button>
+  )
+}
+export default RightArrow
