+++ src/components/Footer.tsx
@@ -1,12 +1,8 @@
 import React from "react"
 
 const Footer = () => (
-  <div>
-    <div>
-      <div className="footer-container">
-        Served by Clarita, the free-software self-hosted web gallery
-      </div>
-    </div>
+  <div className="footer-container">
+    Served by Clarita, the free-software self-hosted web gallery
   </div>
 )
